Add MackerelContainerAgentDefinition and use it internally
