+++ src/add-to-task-definition.test.ts
@@ -8,6 +8,7 @@ describe("addMackerelContainerAgent", () => {
     const taskDefinition = new Ec2TaskDefinition(stack, "TaskDefinition", {})
     const container = addMackerelContainerAgent({
       apiKey: "keep-my-secret",
+      parent: stack,
       taskDefinition,
     })
     expect(stack.toCloudFormation()).toMatchSnapshot()
@@ -18,6 +19,7 @@ describe("addMackerelContainerAgent", () => {
     const taskDefinition = new Ec2TaskDefinition(stack, "TaskDefinition", {})
     const container = addMackerelContainerAgent({
       apiKey: "keep-my-secret",
+      parent: stack,
       roles: [
         { service: "My-service", role: "db" },
         { service: "My-service", role: "proxy" },
@@ -33,6 +35,7 @@ describe("addMackerelContainerAgent", () => {
     const container = addMackerelContainerAgent({
       apiKey: "keep-my-secret",
       ignoreContainer: "(mackerel|xray)",
+      parent: stack,
       taskDefinition,
     })
     expect(stack.toCloudFormation()).toMatchSnapshot()
