+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./add-to-task-definition"
 export * from "./types"
+export * from "./mackerel-container-agent-definition"
